Move component into it’s own directory
