+++ packages/faustwp-core/src/components/FaustAdminBar/index.ts
@@ -1,0 +1,1 @@
+export { FaustAdminBar } from './FaustAdminBar.js';
