+++ packages/faustwp-core/src/components/FaustAdminBar/nodes/Brand.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { adminUrl } from '../../utils/adminUrl.js';
+import { adminUrl } from '../../../utils/adminUrl.js';
 
 const defaultUrl = adminUrl('about.php');
