+++ packages/faustwp-core/src/components/FaustAdminBar/nodes/Comments.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { adminUrl } from '../../utils/adminUrl.js';
+import { adminUrl } from '../../../utils/adminUrl.js';
 
 const defaultUrl = adminUrl('edit-comments.php');
