+++ packages/faustwp-core/src/components/FaustAdminBar/nodes/README.md
