+++ packages/faustwp-core/src/components/FaustAdminBar/nodes/SiteName.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { adminUrl } from '../../utils/adminUrl.js';
+import { adminUrl } from '../../../utils/adminUrl.js';
 
 export function SiteName({ siteName = '' }) {
   const [hover, setHover] = useState(false);
